Split GetLegendGraphic and LiveUpdate (solar term) tests from TestWMS
