+++ tests/AdagucTests/TestWMS.py
@@ -102,22 +102,6 @@ class TestWMS(unittest.TestCase):
             AdagucTestTools().readfromfile(self.expectedoutputsspath + filename),
         )
 
-    def test_WMSGetLegendGraphic_testdatanc(self):
-        AdagucTestTools().cleanTempDir()
-        filename = "test_WMSGetLegendGraphic_testdatanc.png"
-        # pylint: disable=unused-variable
-        status, data, headers = AdagucTestTools().runADAGUCServer(
-            "source=testdata.nc&SERVICE=WMS&VERSION=1.3.0&request=GetLegendGraphic&layer=testdata&format=image/png&STYLE=auto/nearest&layers=testdata&transparent=true",
-            env=self.env,
-            args=["--report"],
-        )
-        AdagucTestTools().writetofile(self.testresultspath + filename, data)
-        self.assertEqual(status, 0)
-        self.assertEqual(
-            data,
-            AdagucTestTools().readfromfile(self.expectedoutputsspath + filename),
-        )
-
     def test_WMSGetMap_Report_env(self):
         AdagucTestTools().cleanTempDir()
         filename = "test_WMSGetMap_Report_env"
@@ -1238,115 +1222,6 @@ class TestWMS(unittest.TestCase):
             AdagucTestTools().readfromfile(self.expectedoutputsspath + filename),
         )
 
-    def test_WMSGetLegendGraphic_adaguc_scaling_dataset_scaling1x(self):
-        AdagucTestTools().cleanTempDir()
-        config = ADAGUC_PATH + "/data/config/adaguc.tests.dataset.xml," + ADAGUC_PATH + "/data/config/datasets/adaguc.tests.scaling.xml"
-        env = {"ADAGUC_CONFIG": config}
-        # pylint: disable=unused-variable
-        status, data, headers = AdagucTestTools().runADAGUCServer(args=["--updatedb", "--config", config], env=self.env, isCGI=False)
-        self.assertEqual(status, 0)
-
-        filename = "test_WMSGetLegendGraphic_adaguc_scaling_dataset_scaling1x.png"
-
-        status, data, headers = AdagucTestTools().runADAGUCServer(
-            "DATASET=adaguc.tests.scaling&SERVICE=WMS&&version=1.1.1&service=WMS&request=GetLegendGraphic&layer=testdata&format=image/png&STYLE=testdata/nearest&layers=testdata&&&transparent=true",
-            env=env,
-        )
-        AdagucTestTools().writetofile(self.testresultspath + filename, data)
-        self.assertEqual(status, 0)
-        self.assertEqual(
-            data,
-            AdagucTestTools().readfromfile(self.expectedoutputsspath + filename),
-        )
-
-    def test_WMSGetMapWithGetLegendGraphic_adaguc_scaling_dataset_scaling1x(self):
-        AdagucTestTools().cleanTempDir()
-        config = ADAGUC_PATH + "/data/config/adaguc.tests.dataset.xml," + ADAGUC_PATH + "/data/config/datasets/adaguc.tests.scaling.xml"
-        env = {"ADAGUC_CONFIG": config}
-        # pylint: disable=unused-variable
-        status, data, headers = AdagucTestTools().runADAGUCServer(args=["--updatedb", "--config", config], env=self.env, isCGI=False)
-        self.assertEqual(status, 0)
-
-        filename = "test_WMSGetMapWithGetLegendGraphic_adaguc_scaling_dataset_scaling1x.png"
-
-        status, data, headers = AdagucTestTools().runADAGUCServer(
-            "DATASET=adaguc.tests.scaling&SERVICE=WMS&&SERVICE=WMS&VERSION=1.3.0&REQUEST=GetMap&LAYERS=testdata&WIDTH=1024&HEIGHT=1024&CRS=EPSG%3A3857&BBOX=-3099408.36963744,3701316.1668297593,3704230.74564144,10633468.46539724&STYLES=testdata%2Fnearest&FORMAT=image/png&TRANSPARENT=TRUE&showlegend=true",
-            env=env,
-        )
-        AdagucTestTools().writetofile(self.testresultspath + filename, data)
-        self.assertEqual(status, 0)
-        self.assertEqual(
-            data,
-            AdagucTestTools().readfromfile(self.expectedoutputsspath + filename),
-        )
-
-    def test_WMSGetMapWithGetLegendGraphic_adaguc_scaling_dataset_scaling4x(self):
-        AdagucTestTools().cleanTempDir()
-        config = ADAGUC_PATH + "/data/config/adaguc.tests.dataset.xml," + ADAGUC_PATH + "/data/config/datasets/adaguc.tests.scaling.xml"
-        env = {"ADAGUC_CONFIG": config}
-        # pylint: disable=unused-variable
-        status, data, headers = AdagucTestTools().runADAGUCServer(args=["--updatedb", "--config", config], env=self.env, isCGI=False)
-        self.assertEqual(status, 0)
-
-        filename = "test_WMSGetMapWithGetLegendGraphic_adaguc_scaling_dataset_scaling4x.png"
-
-        status, data, headers = AdagucTestTools().runADAGUCServer(
-            "DATASET=adaguc.tests.scaling&SERVICE=WMS&&SERVICE=WMS&VERSION=1.3.0&REQUEST=GetMap&LAYERS=testdata&WIDTH=1024&HEIGHT=1024&CRS=EPSG%3A3857&BBOX=-3099408.36963744,3701316.1668297593,3704230.74564144,10633468.46539724&STYLES=testdatascaling4x%2Fnearest&FORMAT=image/png&TRANSPARENT=TRUE&showlegend=true",
-            env=env,
-        )
-        AdagucTestTools().writetofile(self.testresultspath + filename, data)
-        self.assertEqual(status, 0)
-        self.assertEqual(
-            data,
-            AdagucTestTools().readfromfile(self.expectedoutputsspath + filename),
-        )
-
-    def test_WMSGetLegendGraphic_inverted_min_max(self):
-        AdagucTestTools().cleanTempDir()
-        ADAGUC_PATH = os.environ["ADAGUC_PATH"]
-        config = (
-            ADAGUC_PATH + "/data/config/adaguc.tests.dataset.xml," + ADAGUC_PATH + "/data/config/datasets/adaguc.tests.invertedlegend.xml"
-        )
-        env = {"ADAGUC_CONFIG": config}
-
-        status, data, headers = AdagucTestTools().runADAGUCServer(args=["--updatedb", "--config", config], env=self.env, isCGI=False)
-        self.assertEqual(status, 0)
-
-        filename = "test_WMSGetLegendGraphic_inverted_min_max.png"
-        status, data, headers = AdagucTestTools().runADAGUCServer(
-            "DATASET=adaguc.tests.invertedlegend&SERVICE=WMS&&version=1.1.1&service=WMS&request=GetLegendGraphic&layer=testdata&format=image/png&STYLE=testdatainverted/nearest&layers=testdata&&&transparent=true",
-            env=env,
-        )
-        AdagucTestTools().writetofile(self.testresultspath + filename, data)
-        self.assertEqual(status, 0)
-        self.assertEqual(
-            data,
-            AdagucTestTools().readfromfile(self.expectedoutputsspath + filename),
-        )
-
-    def test_WMSGetLegendGraphic_log10(self):
-        AdagucTestTools().cleanTempDir()
-        ADAGUC_PATH = os.environ["ADAGUC_PATH"]
-        config = (
-            ADAGUC_PATH + "/data/config/adaguc.tests.dataset.xml," + ADAGUC_PATH + "/data/config/datasets/adaguc.tests.invertedlegend.xml"
-        )
-        env = {"ADAGUC_CONFIG": config}
-
-        status, data, headers = AdagucTestTools().runADAGUCServer(args=["--updatedb", "--config", config], env=self.env, isCGI=False)
-        self.assertEqual(status, 0)
-
-        filename = "test_WMSGetLegendGraphic_log10.png"
-        status, data, headers = AdagucTestTools().runADAGUCServer(
-            "DATASET=adaguc.tests.invertedlegend&SERVICE=WMS&&version=1.1.1&service=WMS&request=GetLegendGraphic&layer=testdata&format=image/png&STYLE=testdatalog10/nearest&layers=testdata&&&transparent=false",
-            env=env,
-        )
-        AdagucTestTools().writetofile(self.testresultspath + filename, data)
-        self.assertEqual(status, 0)
-        self.assertEqual(
-            data,
-            AdagucTestTools().readfromfile(self.expectedoutputsspath + filename),
-        )
-
     def test_WMSGetMapQuantizeLow(self):
         AdagucTestTools().cleanTempDir()
         ADAGUC_PATH = os.environ["ADAGUC_PATH"]
@@ -1678,31 +1553,6 @@ class TestWMS(unittest.TestCase):
         filename = "test_WMSGetMap_NearestRenderWithShadeIntervalPrecise.png"
         status, data, headers = AdagucTestTools().runADAGUCServer(
             "DATASET=nearestshadeinterval&SERVICE=WMS&SERVICE=WMS&VERSION=1.3.0&REQUEST=GetMap&LAYERS=testdata&WIDTH=256&HEIGHT=256&CRS=EPSG%3A4326&BBOX=30,-30,75,30&STYLES=shadedstyleprecise%2Fnearest&FORMAT=image/png&TRANSPARENT=FALSE&",
-            env=env,
-        )
-        AdagucTestTools().writetofile(self.testresultspath + filename, data)
-        self.assertEqual(status, 0)
-        self.assertEqual(
-            data,
-            AdagucTestTools().readfromfile(self.expectedoutputsspath + filename),
-        )
-
-    def test_WMSGetLegendGraphic_NearestRenderWithShadeInterval(self):
-        AdagucTestTools().cleanTempDir()
-        config = (
-            ADAGUC_PATH
-            + "/data/config/adaguc.tests.dataset.xml,"
-            + ADAGUC_PATH
-            + "/data/config/datasets/adaguc.tests.nearestshadeinterval.xml"
-        )
-        env = {"ADAGUC_CONFIG": config}
-        # pylint: disable=unused-variable
-        status, data, headers = AdagucTestTools().runADAGUCServer(args=["--updatedb", "--config", config], env=self.env, isCGI=False)
-        self.assertEqual(status, 0)
-
-        filename = "test_WMSGetLegendGraphic_NearestRenderWithShadeInterval.png"
-        status, data, headers = AdagucTestTools().runADAGUCServer(
-            "DATASET=nearestshadeinterval&&SERVICE=WMS&VERSION=1.3.0&REQUEST=GetLegendGraphic&LAYER=testdata&WIDTH=150&HEIGHT=800&STYLES=shadedstyleprecise%2Fnearest&FORMAT=image/png&TRANSPARENT=TRUE&",
             env=env,
         )
         AdagucTestTools().writetofile(self.testresultspath + filename, data)
@@ -2433,150 +2283,6 @@ class TestWMS(unittest.TestCase):
         self.assertEqual(status, 0)
         self.assertTrue(AdagucTestTools().compareGetCapabilitiesXML(self.testresultspath + filename, self.expectedoutputsspath + filename))
 
-    def test_WMSGetMap_SolarTerminatorEquinox(self):
-        # Testing the solar terminator on March 21, 2023
-        env = make_adaguc_env("{ADAGUC_PATH}/data/config/datasets/adaguc_tests_solarterminator.xml")
-        update_db(env)
-        filename = "test_WMSGetMap_SolarTerminatorEquinox.png"
-        status, data, headers = AdagucTestTools().runADAGUCServer(
-            "DATASET=solarterminator&SERVICE=WMS&&SERVICE=WMS&VERSION=1.3.0&REQUEST=GetMap&LAYERS=solarterminator&WIDTH=200&HEIGHT=200&CRS=EPSG%3A3857&BBOX=-27591378.677139122,-15819675.465716192,24482445.32432534,23920874.430138264&STYLES=solt_twilight&FORMAT=image/png&TRANSPARENT=TRUE&&time=2023-03-21T00%3A00%3A00Z&",
-            env=env,
-        )
-        AdagucTestTools().writetofile(self.testresultspath + filename, data)
-        self.assertEqual(status, 0)
-        self.assertEqual(
-            data,
-            AdagucTestTools().readfromfile(self.expectedoutputsspath + filename),
-        )
-
-    def test_WMSGetMap_SolarTerminatorSolstice(self):
-        # Testing the solar terminator on December 21, 2022
-        env = make_adaguc_env("{ADAGUC_PATH}/data/config/datasets/adaguc_tests_solarterminator.xml")
-        update_db(env)
-        filename = "test_WMSGetMap_SolarTerminatorSolstice.png"
-        status, data, headers = AdagucTestTools().runADAGUCServer(
-            "DATASET=solarterminator&SERVICE=WMS&&SERVICE=WMS&VERSION=1.3.0&REQUEST=GetMap&LAYERS=solarterminator&WIDTH=200&HEIGHT=200&CRS=EPSG%3A3857&BBOX=-27591378.677139122,-15819675.465716192,24482445.32432534,23920874.430138264&STYLES=solt_twilight&FORMAT=image/png&TRANSPARENT=TRUE&&time=2022-12-21T00%3A00%3A00Z&",
-            env=env,
-        )
-        AdagucTestTools().writetofile(self.testresultspath + filename, data)
-        self.assertEqual(status, 0)
-        self.assertEqual(
-            data,
-            AdagucTestTools().readfromfile(self.expectedoutputsspath + filename),
-        )
-
-    def test_WMSGetMap_SolarTerminatorSolstice_continuous(self):
-        # Testing the solar terminator on December 21, 2022 with continuous colours
-        env = make_adaguc_env("{ADAGUC_PATH}/data/config/datasets/adaguc_tests_solarterminator.xml")
-        update_db(env)
-        filename = "test_WMSGetMap_SolarTerminatorSolstice_continuous.png"
-        status, data, headers = AdagucTestTools().runADAGUCServer(
-            "DATASET=solarterminator&SERVICE=WMS&&SERVICE=WMS&VERSION=1.3.0&REQUEST=GetMap&LAYERS=solarterminator&WIDTH=200&HEIGHT=200&CRS=EPSG%3A3857&BBOX=-27591378.677139122,-15819675.465716192,24482445.32432534,23920874.430138264&STYLES=solt_continuous&FORMAT=image/png&TRANSPARENT=TRUE&&time=2022-12-21T00%3A00%3A00Z&",
-            env=env,
-        )
-        AdagucTestTools().writetofile(self.testresultspath + filename, data)
-        self.assertEqual(status, 0)
-        self.assertEqual(
-            data,
-            AdagucTestTools().readfromfile(self.expectedoutputsspath + filename),
-        )
-
-    def test_WMSGetMap_SolarTerminatorQuarterPoint(self):
-        # Testing the solar terminator on August 7, 2000
-        env = make_adaguc_env("{ADAGUC_PATH}/data/config/datasets/adaguc_tests_solarterminator.xml")
-        update_db(env)
-        filename = "test_WMSGetMap_SolarTerminatorQuarterPoint.png"
-        status, data, headers = AdagucTestTools().runADAGUCServer(
-            "DATASET=solarterminator&SERVICE=WMS&&SERVICE=WMS&VERSION=1.3.0&REQUEST=GetMap&LAYERS=solarterminator&WIDTH=200&HEIGHT=200&CRS=EPSG%3A3857&BBOX=-27591378.677139122,-15819675.465716192,24482445.32432534,23920874.430138264&STYLES=solt_twilight&FORMAT=image/png&TRANSPARENT=TRUE&&time=2000-08-07T00%3A00%3A00Z&",
-            env=env,
-        )
-        AdagucTestTools().writetofile(self.testresultspath + filename, data)
-        self.assertEqual(status, 0)
-        self.assertEqual(
-            data,
-            AdagucTestTools().readfromfile(self.expectedoutputsspath + filename),
-        )
-
-    def test_WMSGetLegendGraphic_SolarTerminator_twilight(self):
-        # Testing the solar terminator legend
-        env = make_adaguc_env("{ADAGUC_PATH}/data/config/datasets/adaguc_tests_solarterminator.xml")
-        update_db(env)
-        filename = "test_WMSGetLegendGraphic_SolarTerminator_twilight.png"
-        status, data, headers = AdagucTestTools().runADAGUCServer(
-            "DATASET=solarterminator&SERVICE=WMS&&version=1.3.0&service=WMS&request=GetLegendGraphic&layer=solarterminator&format=image/png&STYLE=solt_twilight&layers=solarterminator&&&transparent=true&",
-            env=env,
-        )
-        AdagucTestTools().writetofile(self.testresultspath + filename, data)
-        self.assertEqual(status, 0)
-        self.assertEqual(
-            data,
-            AdagucTestTools().readfromfile(self.expectedoutputsspath + filename),
-        )
-
-    def test_WMSGetLegendGraphic_SolarTerminator_continuous(self):
-        # Testing the solar terminator legend
-        env = make_adaguc_env("{ADAGUC_PATH}/data/config/datasets/adaguc_tests_solarterminator.xml")
-        update_db(env)
-        filename = "test_WMSGetLegendGraphic_SolarTerminator_continuous.png"
-        status, data, headers = AdagucTestTools().runADAGUCServer(
-            "DATASET=solarterminator&SERVICE=WMS&&version=1.3.0&service=WMS&request=GetLegendGraphic&layer=solarterminator&format=image/png&STYLE=solt_continuous&layers=solarterminator&&&transparent=true&",
-            env=env,
-        )
-        AdagucTestTools().writetofile(self.testresultspath + filename, data)
-        self.assertEqual(status, 0)
-        self.assertEqual(
-            data,
-            AdagucTestTools().readfromfile(self.expectedoutputsspath + filename),
-        )
-
-    def test_WMSGetFeatureInfo_SolarTerminator(self):
-        # Testing the solar terminator feature info
-        env = make_adaguc_env("{ADAGUC_PATH}/data/config/datasets/adaguc_tests_solarterminator.xml")
-        update_db(env)
-        filename = "test_WMSGetFeatureInfo_SolarTerminator.html"
-        status, data, headers = AdagucTestTools().runADAGUCServer(
-            "dataset=adaguc_tests_solarterminator&&SERVICE=WMS&REQUEST=GetFeatureInfo&VERSION=1.3.0&LAYERS=solarterminator&QUERY_LAYERS=solarterminator&CRS=EPSG%3A3857&BBOX=-48040284.36018957,-50974535.88541803,-10040284.360189572,62500602.23612894&WIDTH=362&HEIGHT=1081&I=354&J=481&INFO_FORMAT=text/html&STYLES=&&time=2026-01-05T16%3A40%3A00Z&",
-            env=env,
-        )
-        AdagucTestTools().writetofile(self.testresultspath + filename, data)
-        self.assertEqual(status, 0)
-        self.assertEqual(
-            data,
-            AdagucTestTools().readfromfile(self.expectedoutputsspath + filename),
-        )
-
-    def test_WMSGetFeatureInfo_timeseries_SolarTerminator(self):
-        # Testing the solar terminator feature info
-        env = make_adaguc_env("{ADAGUC_PATH}/data/config/datasets/adaguc_tests_solarterminator.xml")
-        update_db(env)
-        filename = "test_WMSGetFeatureInfo_timeseries_SolarTerminator.html"
-        status, data, headers = AdagucTestTools().runADAGUCServer(
-            "dataset=solt&service=WMS&request=GetFeatureInfo&version=1.3.0&layers=solarterminator&query_layers=solarterminator&crs=EPSG%3A3857&bbox=-19000000%2C-78286307.05394192%2C19000000%2C78286307.05394192&width=241&height=993&i=180&j=366&format=image%2Fgif&info_format=application%2Fjson&time=2025-10-01T00%3A00%3A00Z%2F2025-10-02T00%3A00%3A00Z&",
-            env=env,
-        )
-        AdagucTestTools().writetofile(self.testresultspath + filename, data)
-        self.assertEqual(status, 0)
-        self.assertEqual(
-            data,
-            AdagucTestTools().readfromfile(self.expectedoutputsspath + filename),
-        )
-
-    def test_WMSGetCapabilities_SolarTerminator(self):
-        # Testing the solar terminator capabilities
-        env = make_adaguc_env("{ADAGUC_PATH}/data/config/datasets/adaguc.tests.solarterminator-custom.xml")
-        update_db(env)
-        filename = "test_WMSGetCapabilities_SolarTerminator.xml"
-        status, data, headers = AdagucTestTools().runADAGUCServer(
-            "dataset=solt&SERVICE=WMS&request=GetCapabilities",
-            env=env,
-        )
-        AdagucTestTools().writetofile(self.testresultspath + filename, data)
-        self.assertEqual(status, 0)
-
-        AdagucTestTools().writetofile(self.testresultspath + filename, data)
-        self.assertEqual(status, 0)
-        self.assertTrue(AdagucTestTools().compareTimeRange(self.testresultspath + filename, 7, "PT1H", "solarterminator"))
-
     def test_WMSGetCapabilities_403_default_time_referenced_to_forecast_time(self):
         """
         See https://github.com/KNMI/adaguc-server/issues/403
@@ -2602,114 +2308,6 @@ class TestWMS(unittest.TestCase):
         AdagucTestTools().writetofile(self.testresultspath + filename, data)
         self.assertEqual(status, 0)
         self.assertTrue(AdagucTestTools().compareGetCapabilitiesXML(self.testresultspath + filename, self.expectedoutputsspath + filename))
-
-    def test_WMSGetLegendGraphic_LongTempLegend(self):
-        """Skip some labels when legend is too long"""
-        """Corresponds to the special case for definedLegendOnShadeClasses case in CCreateLegendRenderDiscreteLegend """
-        AdagucTestTools().cleanTempDir()
-        config = (
-            ADAGUC_PATH
-            + "/data/config/adaguc.tests.dataset.xml,"
-            + ADAGUC_PATH
-            + "/data/config/datasets/test.uwcw_ha43_dini_5p5km_10x8.xml"
-        )
-        env = {"ADAGUC_CONFIG": config}
-        # pylint: disable=unused-variable
-        status, data, headers = AdagucTestTools().runADAGUCServer(args=["--updatedb", "--config", config], env=self.env, isCGI=False)
-        self.assertEqual(status, 0)
-
-        filename = "test_WMSGetLegendGraphic_LongTempLegend.png"
-        status, data, headers = AdagucTestTools().runADAGUCServer(
-            "DATASET=test.uwcw_ha43_dini_5p5km_10x8&&SERVICE=WMS&VERSION=1.3.0&REQUEST=GetLegendGraphic&LAYER=air_temperature_pl&WIDTH=100&HEIGHT=400&STYLE=temperature%2Fshaded&FORMAT=image/png&TRANSPARENT=TRUE&",
-            env=env,
-        )
-        AdagucTestTools().writetofile(self.testresultspath + filename, data)
-        self.assertEqual(status, 0)
-        self.assertEqual(
-            data,
-            AdagucTestTools().readfromfile(self.expectedoutputsspath + filename),
-        )
-
-    def test_WMSGetLegendGraphic_LongTempLegend_height260(self):
-        """Skip some labels when legend is too long, making sure the whole range is contained"""
-        """Corresponds to the special case for definedLegendOnShadeClasses case in CCreateLegendRenderDiscreteLegend """
-        AdagucTestTools().cleanTempDir()
-        config = (
-            ADAGUC_PATH
-            + "/data/config/adaguc.tests.dataset.xml,"
-            + ADAGUC_PATH
-            + "/data/config/datasets/test.uwcw_ha43_dini_5p5km_10x8.xml"
-        )
-        env = {"ADAGUC_CONFIG": config}
-        # pylint: disable=unused-variable
-        status, data, headers = AdagucTestTools().runADAGUCServer(args=["--updatedb", "--config", config], env=self.env, isCGI=False)
-        self.assertEqual(status, 0)
-
-        filename = "test_WMSGetLegendGraphic_LongTempLegend_height260.png"
-        status, data, headers = AdagucTestTools().runADAGUCServer(
-            "DATASET=test.uwcw_ha43_dini_5p5km_10x8&&SERVICE=WMS&VERSION=1.3.0&REQUEST=GetLegendGraphic&LAYER=air_temperature_pl&WIDTH=260&HEIGHT=260&STYLE=temperature%2Fshaded&FORMAT=image/png&TRANSPARENT=TRUE&",
-            env=env,
-        )
-        AdagucTestTools().writetofile(self.testresultspath + filename, data)
-        self.assertEqual(status, 0)
-        self.assertEqual(
-            data,
-            AdagucTestTools().readfromfile(self.expectedoutputsspath + filename),
-        )
-
-    def test_WMSGetLegendGraphic_LongTempLegend_clipping(self):
-        """Skip some labels when legend is too long and clip"""
-        """Corresponds to the special case for definedLegendOnShadeClasses case in CCreateLegendRenderDiscreteLegend """
-        AdagucTestTools().cleanTempDir()
-        config = (
-            ADAGUC_PATH
-            + "/data/config/adaguc.tests.dataset.xml,"
-            + ADAGUC_PATH
-            + "/data/config/datasets/test.uwcw_ha43_dini_5p5km_10x8.xml"
-        )
-        env = {"ADAGUC_CONFIG": config}
-        # pylint: disable=unused-variable
-        status, data, headers = AdagucTestTools().runADAGUCServer(args=["--updatedb", "--config", config], env=self.env, isCGI=False)
-        self.assertEqual(status, 0)
-
-        filename = "test_WMSGetLegendGraphic_LongTempLegend_clipping.png"
-        status, data, headers = AdagucTestTools().runADAGUCServer(
-            "DATASET=test.uwcw_ha43_dini_5p5km_10x8&&SERVICE=WMS&VERSION=1.3.0&REQUEST=GetLegendGraphic&LAYER=air_temperature_pl&WIDTH=100&HEIGHT=400&STYLE=temperature_clip%2Fshaded&FORMAT=image/png&TRANSPARENT=TRUE&",
-            env=env,
-        )
-        AdagucTestTools().writetofile(self.testresultspath + filename, data)
-        self.assertEqual(status, 0)
-        self.assertEqual(
-            data,
-            AdagucTestTools().readfromfile(self.expectedoutputsspath + filename),
-        )
-
-    def test_WMSGetLegendGraphic_LongTempLegend2(self):
-        """Another test showing another type of legend that skips classes without data"""
-        """Corresponds to the discreteLegendOnInterval case in CCreateLegendRenderDiscreteLegend """
-        AdagucTestTools().cleanTempDir()
-        config = (
-            ADAGUC_PATH
-            + "/data/config/adaguc.tests.dataset.xml,"
-            + ADAGUC_PATH
-            + "/data/config/datasets/test.uwcw_ha43_dini_5p5km_10x8.xml"
-        )
-        env = {"ADAGUC_CONFIG": config}
-        # pylint: disable=unused-variable
-        status, data, headers = AdagucTestTools().runADAGUCServer(args=["--updatedb", "--config", config], env=self.env, isCGI=False)
-        self.assertEqual(status, 0)
-
-        filename = "test_WMSGetLegendGraphic_LongTempLegend2.png"
-        status, data, headers = AdagucTestTools().runADAGUCServer(
-            "DATASET=test.uwcw_ha43_dini_5p5km_10x8&SERVICE=WMS&&version=1.3.0&service=WMS&request=GetLegendGraphic&layer=air_temperature_pl&format=image/png&STYLE=temperature_wow/shaded&TRANSPARENT=TRUE&",
-            env=env,
-        )
-        AdagucTestTools().writetofile(self.testresultspath + filename, data)
-        self.assertEqual(status, 0)
-        self.assertEqual(
-            data,
-            AdagucTestTools().readfromfile(self.expectedoutputsspath + filename),
-        )
 
     def test_WMS_MultiDim_PreconfiguredDataBaseTable(self):
         """A dataset configuration with multiple dimensions and a DataBaseTable should work"""
