+++ tests/expectedoutputs/TestWMSGetLegendGraphic/test_WMSGetLegendGraphic_LongTempLegend.png
