+++ tests/expectedoutputs/TestWMSGetLegendGraphic/test_WMSGetLegendGraphic_LongTempLegend_height260.png
