+++ tests/expectedoutputs/TestWMSGetLegendGraphic/test_WMSGetLegendGraphic_NearestRenderWithShadeInterval.png
