+++ tests/expectedoutputs/TestWMSGetLegendGraphic/test_WMSGetLegendGraphic_adaguc_scaling_dataset_scaling1x.png
