+++ tests/expectedoutputs/TestWMSGetLegendGraphic/test_WMSGetLegendGraphic_inverted_min_max.png
