+++ tests/expectedoutputs/TestWMSGetLegendGraphic/test_WMSGetMapWithGetLegendGraphic_adaguc_scaling_dataset_scaling4x.png
