+++ tests/expectedoutputs/TestWMSLiveUpdate/test_WMSLiveUpdateGetFeatureInfo_SolarTerminator.html
