+++ tests/expectedoutputs/TestWMSLiveUpdate/test_WMSLiveUpdateGetFeatureInfo_timeseries_SolarTerminator.html
