+++ tests/expectedoutputs/TestWMSLiveUpdate/test_WMSLiveUpdateGetLegendGraphic_SolarTerminator_continuous.png
