+++ tests/expectedoutputs/TestWMSLiveUpdate/test_WMSLiveUpdateGetMap_SolarTerminatorEquinox.png
