+++ tests/expectedoutputs/TestWMSLiveUpdate/test_WMSLiveUpdateGetMap_SolarTerminatorQuarterPoint.png
