+++ tests/expectedoutputs/TestWMSLiveUpdate/test_WMSLiveUpdateGetMap_SolarTerminatorSolstice.png
