+++ tests/expectedoutputs/TestWMSLiveUpdate/test_WMSLiveUpdateGetMap_SolarTerminatorSolstice_continuous.png
